Rename files using kebab style
