+++ src/rectangle-tool.js
@@ -4,7 +4,7 @@ import { styleMap } from 'lit-html/directives/style-map.js';
 
 import Merge from './merge';
 import Utils from './utils';
-import BaseTool from './basetool';
+import BaseTool from './base-tool';
 
 /** ****************************************************************************
   * RectangleTool class
